+++ wls-gui-wahllokalsystem/src/composables/tasks/taskFactories/kopfdatenTaskFactory.ts
