+++ wls-gui-wahllokalsystem/src/composables/tasks/taskFactories/ungueltigeWahlscheineTaskFactory.ts
