+++ wls-gui-wahllokalsystem/src/composables/tasks/taskFactories/wahlvorstandTaskFactory.ts
